requirements.txt
asgiref==3.6.0
Django==4.1.5
django-autoslug==1.9.8
Pillow==9.4.0
python-dotenv==0.21.0
sqlparse==0.4.3
